requirements.txt
requests==2.26.0
pycoingecko==3.1.0
python-dateutil==2.8.2
